Re-calculate alarms when computer is back from sleep
Since Javascript does not provide an API to fire events at given times, to fire a
browser notification for a calendar event we set a timeout with the appropriate delay.
If the computer is suspended, for Javascript no time is passing, meaning the timeout will
trigger later than expected. If your computer slept for 8 hours and you didn't reload your tab,
your notification would arrive 8 hours later, which is not great.
This MR introduces a hook that allows to detect when the computer goes to sleep, and forces
a timeout delay re-computation when the computer wakes up.

Additionally, we recently changed the lookahead time with poor judgement. With the new lookahead of 12 hours
the typical user would load the app in the morning, basically fetching the alarms for that day only,
then fetch again late in the day the alarms for the night (probably none).
If the computer then went to sleep say 8 hours in the night, before fetching alarms for the next day, the
next fetch would be delayed 8 hours and instead of fetching in the morning again, the fetch would happen
in the late afternoon, missing all the alarms in the day :(

CALWEB-2621