Handle alarm trigger strings for full-day events generated by other clients
If you want a notification for a full day event that triggers 8 days before at a certain time during the day.
When generating such a notification from the web client, the trigger string looks like -P7DT4H.
An equivalent trigger string is -P1WT4H. Both should be parsed equivalently, but instead we were interpreting
the latter as a notification that should happen "1 week before", cropping 1 day. Both the API and other clients
understand the -P1WT4H notation, so we should as well.

An easy fix is to re-use an existing normalization helper that transforms -P1WT4H into -P7DT4H before
parsing the trigger in the notification model

CALWEB-2718