Fix target environment condition
There was a bug where if your cookie-at-load was "alpha" but your
early-access-scope "beta" you wouldn't get the desynchronized banner
showing up. This happened because the "getTargetEnvironment" function
would return the initially passed cookie if the cookie was both defined
and valid, however this didn't account for the possibility that the cookie
would be different from the early access scope setting, a condition
which would slip through this test:
- cookie is defined
- cookie is valid (meaning that it is included in the possible options)
- cookie is NOT of the value set as the early access scope

Simplified the condition to either return undefined or the value of the
early access scope. This consequently got rid of a bunch of additional
code that dealth with tracking the value of the cookie as it is being
changed in runtime. Since this data was only used in the
getTargetEnvironment function, the code for computing / tracking it
is not necessary any longer.