Set default border-radius to 3
This is the same value as was previously assigned to the SASS
variable $global-border-radius.

While we can't just yet actually propagate the border-radius
changes for real, this commit together with the previous ones
relating to the border-radius update (CP-2422) are setting the
system in place to later transition at the change of a single
variable.

These changes will not be fully invisible but nearly so. Some
border-radiuses have been re-considered, also the old system
with just $border-radius-global & $border-bigger-radius-global
is not fully compatible with the new system (small, medium, large).

CP-2422