Control subscribe to calendar feature via BE feature flag. Improve not synced error message for users
While in beta, we want to be able to disable server-side the "subscribe to calendar". We use a BE feature flag to control that.
If the feature flag is on, the "Subscribe to calendar" button will appear disabled with a tooltip "This feature is unavailable
for the moment".
We also improve the user experience when facing the "not synced" message, by providing more details via tooltip by translating
different BE statuses in more user-friendly error messages.

CALWEB-2807