Recovery settings restructure with recovery file
Contains a rework of the recovery settings and includes the new
'Recovery file' recovery method.

- Renames "Security" page to "Password & security", and
'Passord & recovery' page to 'Recovery'

- Moves "Password", and "Delete" sections into "Password & security"

- Adds an overview section to the "Recovery" page which outlines the
current state of recovery setup

- Adds notification dot to "Recovery" sidebar list item which shows
if either the account recovery or data recovery setup is incomplete

- Add notification dot to user dropdown which shows if either the
account recovery or data recovery setup is incomplete AND the user has
not visited the recovery page

- When notification dot in user dropdown is shown, there is an
additional item in the dropdown which switches between 2 states
depending on whether the mnemonic is enabled.

- Adds export recovery file button to the "Data recovery" section.
The recovery file is an encrypted PGP message containing all active
user keys and is encrypted with a secret stored on the server.

- Adds file recovery method to reactivate keys modal

- Story book updates for the "Card" and "Collapsible" components

CP-2523

Co-authored-by: Mattias Svanström <mmso@mmso.se>
Co-authored-by: Nicolas Hoffmann <dante3333@gmail.com>
Co-authored-by: Daniel Huigens <daniel.huigens@proton.ch>