Fix Calendar icon condition on collapsed messages with attachment
When a collapsed message had non ICS attachments, a Calendar icon was showing instead of an attachment icon. This PR fixes that problem.

However in order to show the calendar icon correctly on collapsed messages, attachments list should be provided by the backend API.

CALWEB-2717

Add mention of the feature in the mail changelog