Fix React dependency that was making subscribed calendars to be labeled as not synced
In the useSubscribedCalendars hook, there was an useEffect that would run on every change of the
calendars reference (in particular on every calendars update). This hook would set the subscription
parameters of all subscribed calendars to whatever their value was in the first run of the hook as
the getCalendarSubscription is cached from the first run. The useEffect should be triggered just on
the component mount, as changes afterwards are propagated via the event loop.

Additionally, the useGetCalendarSubscription hook has been removed as it was a bit dangerous to have
caching in the subscription which could be easily misused as here.

CALWEB-2481