Add padding to embedded verification app
Initially we wanted to let the embedded decide padding, however Android
pointed out that if the embedder added padding and the content inside
the web view became scrollable, the scrollbar would be inside the
padding and not at the very edge of the whole construct of the embedded
hv app. Therefore, even if opinionated about usage, padding inside.

CP-2489