Copy updates & interaction  disabling of completed items
Change completed checklist items to, besides being greyed out and
strikethrough, disallow a user to click them, remove the hover
interaction as well as the angle icon.

Not all of these actions are repeatable after having been done
once initially (sending a message is repeatable, but setting up a
mnemonic recovery for example is not repeatable if already set).
Since the frontend by design has very little knowledge about what
completes an item and the backend is in control of that, we have to
either disallow these items from being interactable when completed
or we have to duplicate the knowledge about item completion in the
frontend which is highly undesireable.

CP-2526