Show different incentive text based on expiry
A note on the wording "Expires soon" even though the checklist is
already expired from a technical perspective.

There are use-cases where even though technically expired, the backend
does not yet remove the get started checklist for a particular user but
rather allows them to still complete the checklist items even after expiry.

It is my assumption that the rationale behind showing "Expires soon" instead
of showing something like "Expired" is that a user might think that "Expired"
meant that he could no longer complete the checklist which is not the case
under certain circumstances.

CP-2526