Catch properly errors in helpers generating occurrences of recurring events
We were already catching potential errors thrown by ICAL.js in these helpers, but it turns out the errors could be thrown
a bit earlier than the try... catch block. For instance, a recurrence rule with both a DAILY/WEEKLY/MONTHLY FREQ part and
a BYYEARDAY rule, which is forbidden by the RFC, would throw earlier at the computation of the iterator.

CALWEB-3354