Add some rudimentary tests to Slider
Unfortunately testing-library doesn't physically render the html inside
its virtual dom, resulting in the elements not having dimensions.
Therefore getBoundingClientRect doesn't work in this environment and
it's a bit tricky to organically test this component via touch & mouse
moves that move a certain distance in x direction.

Have to investigate how to mock this properly, but it doesn't seem trivial,
so for now, gonna add these bunch of smoke tests.