Remove DeleteAccountModal delete button disabled prop
When the modal height is small, the checkbox is not shown. In this
case, some users were complaining that they could not delete their
account because the delete button was disabled.

I have removed the disabled prop, falling back on the native required
prop of the Checkbox resulting in the browser showing a prompt if
the checkbox has not been checked.

CP-3828