Fix image sender border
A border was appearing around the sender image, because there was a 1px transparent border on the container.
As I couldn't know on the container what was the status of the image inside, I just decided to remove this border and make the element 32px. This border can't appear now, because the container is the same size as the image. I've checked the focus-visible state, no impact.

- make element go to 32px
- remove transparent border

MAILWEB-3680
UXE-148