Minimise singup layout shift
Layout shift was being caused by the subtitle suddenly appearing.
This is fixed by assuming there will be a subtitle, and filing
the space with a zero width space character.

Currently, a layout shift may still occur if the signupType is `Email`,
howver the effect is minimised by shifting the layout when the inputs
appear.

Another cause is the external account switch. This has been minimised
by displaying when the inputs are shown.