Fix issue in styles order
Calendar QA spotted an issue in share modal autocomplete, a spacing issue.

This issue is caused by moving input to atoms: a79cec765fc7421eb69699db3025afbbab6dacc7

The issue is that by doing this, style of atoms are loaded AFTER helpers (thus getting priority on it), wheareas
it should not. And it's pretty bad.

As a temporary workaround discussed with @msvanstr, we decided to move back the input styles into field-two
ones, until we found a more reliable solution.

Co-authored-by: @msvanstr

UXE-367