Rework onboarding modals and login layout
The login has changed a bit, the loading when creating an account is much simpler and there is no checkbox when selecting a recovery method.
Besides that, the onboarding modal for mail have changed, some steps were deleted and other added to better present Proton value. All the mail modals have a fixed height now, to avoid the content to jump when the user switch between modals.
There is also a new Props for `OnboardingModal` to add steps after the classical steps for more complex onboarding modals flow