Fix delete modal
Delete modal were putting the folder/label/filter name in the title, so when it's long, display is not good.

On top of this, designers/copy teams don't want us to do it.

- removed user-generated content from title
- wrapped them with ellipsis in the content
- small update of copy for folder, given by copy team in a discussion.

Please note: I did not use quotes (even if a grammarNazi like me would love them), because of this:
- in case of super long names it will give break after the quote opening
- and given the complexity of localizing quotes (https://en.wikipedia.org/wiki/Quotation_mark#Summary_table), I went for something simpler (I can't predict the length of quotes, for some language, it might be longer).

UXE-801